refactor(workflow): split worktree functions to separate module
- Extract createWorktree, createWorktreeFromRepo, listWorktrees,
  removeWorktree to git-workflow-worktree.ts
- Reduces git-workflow.ts from 433 to 242 LOC (under 300 limit)
- New git-workflow-worktree.ts is 197 LOC
- Add warning logs for fetch/tracking failures (previously silent)
- Consume stderr in getGitDir() to prevent pipe issues
- Re-export worktree functions from git-workflow.ts for compatibility

Addresses code review findings:
- CLAUDE.md file LOC limit compliance
- Silent failure detection improvements
